+++ esp-hal/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `send_break` for sending software breaks with the UART driver (#3177)
+
 ### Changed
 
 ### Fixed